+++ module/applications/compendium/tabs/equipment.mjs
@@ -44,7 +44,7 @@ export class CompendiumBrowserEquipmentTab extends CompendiumBrowserTab {
       };
       const price = {
         default: "system.price.value",
-        starshipmod: "system.basecost.value"
+        starshipmod: "system.baseCost.value"
       };
       const properties = {
         weapon: "system.properties",
@@ -60,9 +60,11 @@ export class CompendiumBrowserEquipmentTab extends CompendiumBrowserTab {
       };
       const subcategory = {
         consumable: "system.ammoType",
-        weapon: "system.weaponClass"
+        starshipmod: "system.grade.value",
+        weapon: "system.weaponClass",
       };
       const optional = {
+        starshipmod: ["system.baseCost.value", "system.grade.value"],
         weapon: ["system.weaponClass"]
       };
       const indexFields = [...new Set([
@@ -104,9 +106,20 @@ export class CompendiumBrowserEquipmentTab extends CompendiumBrowserTab {
             // Add properties into the correct format for filtering
             if (itemData?.system?.itemTypes?.value === undefined) itemData.system.itemTypes = { value: itemData.type };
             if (itemData?.system?.rarity?.value === undefined) itemData.system.rarity = { value: itemData.system.rarity };
-            if (itemData?.system?.source?.value === undefined) itemData.system.source = { value: itemData.system.source };
-            if (itemData?.system?.category?.value === undefined) itemData.system.category = { value: (itemData.type in category) ? foundry.utils.getProperty(itemData, category[itemData.type]) : null };
-            if (itemData?.system?.subcategory?.value === undefined) itemData.system.subcategory = { value: (itemData.type in subcategory) ? foundry.utils.getProperty(itemData, subcategory[itemData.type]) : null };
+            if (itemData?.system?.source?.value === undefined) itemData.system.source = {
+              value: itemData.system.source.label ?? itemData.system.source.custom ?? itemData.system.source
+            };
+            if (itemData?.system?.category?.value === undefined) itemData.system.category = {
+              value: (itemData.type in category) ? foundry.utils.getProperty(itemData, category[itemData.type]) : null
+            };
+            if (itemData?.system?.subcategory?.value === undefined) itemData.system.subcategory = {
+              value: (itemData.type in subcategory) ? foundry.utils.getProperty(itemData, subcategory[itemData.type]) : null
+            };
+            if (itemData.type === "starshipmod" && [null,undefined].includes(itemData.system.baseCost?.value)) {
+              const baseCost = CONFIG.SW5E.ssModSystemsBaseCost[itemData.system.system.value.toLowerCase()];
+              const gradeMult = Number.isNumeric(itemData.system.grade?.value) ? itemData.system.grade?.value || 1 : 1;
+              itemData.system.baseCost = { value: baseCost * gradeMult };
+            }
             itemData.filters = {};
 
             // Prepare source