+++ module/canvas/_module.mjs
@@ -1,4 +1,5 @@
 export { default as AbilityTemplate } from "./ability-template.mjs";
+export { default as FiringArcTemplate } from "./firing-arc-template.mjs";
 export * as detectionModes from "./detection-modes/_module.mjs";
 export { default as Token5e } from "./token.mjs";
 export { measureDistances } from "./grid.mjs";