+++ module/config.mjs
@@ -127,7 +127,7 @@ SW5E.shieldPointsAbility = "str";
  */
 
 /**
- * The set of skill which can be trained with their default ability scores.
+ * The set of skills which can be trained by characters with their default ability scores.
  * @enum {SkillConfiguration}
  */
 SW5E.skills = {
@@ -153,7 +153,7 @@ SW5E.skills = {
 preLocalize("skills", { key: "label", sort: true });
 
 /**
- * The set of skill which can be trained on starships with their default ability scores.
+ * The set of skills which can be trained on starships with their default ability scores.
  * @enum {SkillConfiguration}
  */
 SW5E.starshipSkills = {
@@ -173,6 +173,15 @@ SW5E.starshipSkills = {
   swn: { label: "SW5E.StarshipSkillSwn", ability: "cha", fullKey: "swindle" }
 };
 preLocalize("starshipSkills", { key: "label", sort: true });
+
+/**
+ * The set of skills which can be trained by characters or starships with their default ability scores.
+ * @enum {SkillConfiguration}
+ */
+SW5E.allSkills = {
+  ...SW5E.skills,
+  ...SW5E.starshipSkills
+};
 
 /* -------------------------------------------- */
 
@@ -2381,6 +2390,46 @@ SW5E.weaponClasses = {
 };
 preLocalize("weaponClasses");
 
+/**
+ * The set of firing arcs for starship weapons.
+ * @enum {{
+ *    label: string,
+ *    direction: [number],
+ *    angle: [number],
+ * }}
+ */
+SW5E.weaponFiringArcs = {
+  forward: {
+    label: "SW5E.FiringArcForward",
+    direction: 0,
+    shape: "cone"
+  },
+  rear: {
+    label: "SW5E.FiringArcRear",
+    direction: 180,
+    shape: "cone"
+  },
+  left: {
+    label: "SW5E.FiringArcLeft",
+    direction: 270,
+    shape: "cone"
+  },
+  right: {
+    label: "SW5E.FiringArcRight",
+    direction: 90,
+    shape: "cone"
+  },
+  turret: {
+    label: "SW5E.FiringArcTurret",
+    direction: 0,
+    shape: "circle"
+  },
+  broadside: {
+    label: "SW5E.FiringArcBroadside"
+  }
+};
+preLocalize("weaponFiringArcs", { keys: ["label"], sort: false });
+
 /* -------------------------------------------- */
 
 /**
@@ -3996,7 +4045,7 @@ SW5E.midiFlags = {
     midiClone: true,
     type: Boolean
   },
-  ...Object.fromEntries(Object.keys(SW5E.skills).map((key => [
+  ...Object.fromEntries(Object.keys(SW5E.allSkills).map((key => [
     `advantage.skill.${key}`,
     {
       name: `SW5E.CharacterFlags.Advantage.Skill.${key}.Name`,
@@ -4340,13 +4389,12 @@ Object.defineProperty(SW5E, "enrichmentLookup", {
   get() {
     if ( !_enrichmentLookup ) {
       _enrichmentLookup = {
-        abilities: {},
-        skills: {},
-        tools: {}
+        abilities: foundry.utils.deepClone(SW5E.abilities),
+        skills: foundry.utils.deepClone(SW5E.skills),
+        tools: foundry.utils.deepClone(SW5E.toolIds)
       };
-      Object.entries(SW5E.abilities).forEach(([k, a]) => _enrichmentLookup.abilities[k] = _enrichmentLookup.abilities[a.fullKey] = a);
-      Object.entries(SW5E.skills).forEach(([k, s]) => _enrichmentLookup.skills[k] = _enrichmentLookup.skills[s.fullKey] = s);
-      Object.entries(SW5E.toolIds).forEach(([k, t]) => _enrichmentLookup.tools[k] = t);
+      Object.entries(SW5E.abilities).forEach(([k, a]) => _enrichmentLookup.abilities[a.fullKey] = { ...a, key: k });
+      Object.entries(SW5E.skills).forEach(([k, s]) => _enrichmentLookup.skills[s.fullKey] = { ...s, key: k });
     }
     return _enrichmentLookup;
   },