+++ module/documents/actor/select-choices.mjs
@@ -64,6 +64,26 @@ export default class SelectChoices {
   /* -------------------------------------------- */
 
   /**
+   * Find key and value for the provided key or key suffix.
+   * @param {string} key  Full prefixed key (e.g. `tool:art:alchemist`) or just the suffix (e.g. `alchemist`).
+   * @returns {[string, SelectChoicesEntry]|null}  An array with the first value being the matched key,
+   *                                               and the second being the value.
+   */
+  find(key) {
+    for ( const [k, v] of Object.entries(this) ) {
+      if ( (k === key) || k.endsWith(`:${key}`) ) {
+        return [k, v];
+      } else if ( v.children ) {
+        const result = v.children.find(key);
+        if ( result ) return result;
+      }
+    }
+    return null;
+  }
+
+  /* -------------------------------------------- */
+
+  /**
    * Merge another SelectChoices object into this one.
    * @param {SelectChoices} other
    * @param {object} [options={}]